backend/requirements.txt
django
django-ninja
django-cors-headers
tree-sitter
tree-sitter-python
langchain
langchain-openai
langchain-community
psycopg2-binary
python-dotenv